requirements.txt
Django==3.1
django-bootstrap-form==3.4
djangorestframework==3.11.1
Pillow==7.2.0
pkg-resources==0.0.0
pylint-django==2.3.0
wrapt==1.12.1
